moved colors file to theme folder
