+++ src/components/rounded-box/rounded-list-box.tsx
@@ -1,8 +1,8 @@
 import { ListItem, UnorderedList } from '@chakra-ui/react';
 import type { FC } from 'react';
 import ReactMarkdown from 'react-markdown';
-import { colors } from '~/common/colors';
 import { RoundedBox } from '~/components/rounded-box/rounded-box';
+import { colors } from '~/theme/colors';
 
 type RoundedListBoxProps = {
 	list: string[];
