+++ src/theme/components.ts
@@ -1,4 +1,4 @@
-import { colors } from '~/common/colors';
+import { colors } from '~/theme/colors';
 
 const navButtonStyle = {
 	textTransform: 'uppercase',
