+++ src/theme/index.ts
@@ -1,5 +1,5 @@
 import { extendTheme } from '@chakra-ui/react';
-import { colors } from '~/common/colors';
+import { colors } from '~/theme/colors';
 import { components } from '~/theme/components';
 import { fonts } from '~/theme/fonts';
